Moving SecurityQualityOfService to Interop namespace, migrating usages
